Feature/ode 651 hm

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetLLB22Builder.java
@@ -8,7 +8,7 @@ public class OffsetLLB22Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetLLB22(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(7).intValue();
+   public static Long offsetLLB22(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(7).longValue();
    }
 }
